\name{daggrForest-package}
\alias{daggrForest-package}
\alias{daggrForest}
\docType{package}
\title{
Downsampled and aggregated random forests for unbalanced class prediction 
}
\description{
Downsampled and aggregated random forests for unbalanced class prediction }
\details{
\tabular{ll}{
Package: \tab daggrForest\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-18\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Subhabrata Majumdar

Maintainer: Subhabrata Majumdar <majum010@umn.edu>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[randomForest:randomForest-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
